initial commit: copy contracts from rainbow-bridge
Copy contracts as-is from github.com/near/rainbow-bridge. This means
they contain invalid imports. The files they import will need to be made
available as libraries, but this has not yet been done.

This also copies the sequence diagram from the `docs` repo of
near/rainbow-bridge into the README and adds extra info about how this
repository fits into the Rainbow Bridge, how to use it, and how to
deploy it.